refactor: extract metric dialogs into reusable components with shared base
