+++ src/app/metric/_components/MetricDialogBase.tsx
@@ -1,0 +1,123 @@
+"use client";
+
+import { useState } from "react";
+
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger,
+} from "@/components/ui/dialog";
+import { api } from "@/trpc/react";
+
+export interface MetricCreateInput {
+  templateId: string;
+  connectionId: string;
+  name: string;
+  description: string;
+  endpointParams: Record<string, string>;
+}
+
+export interface ContentProps {
+  connection: {
+    id: string;
+    connectionId: string;
+    integrationId: string;
+    createdAt: Date;
+  };
+  onSubmit: (data: MetricCreateInput) => void;
+  isCreating: boolean;
+  error: string | null;
+}
+
+interface MetricDialogBaseProps {
+  integrationId: string;
+  title: string;
+  description?: string;
+  trigger?: React.ReactNode;
+  open?: boolean;
+  onOpenChange?: (open: boolean) => void;
+  onSuccess?: () => void;
+  children: (props: ContentProps) => React.ReactNode;
+}
+
+export function MetricDialogBase({
+  integrationId,
+  title,
+  description,
+  trigger,
+  open: controlledOpen,
+  onOpenChange,
+  onSuccess,
+  children,
+}: MetricDialogBaseProps) {
+  const [internalOpen, setInternalOpen] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+
+  const isControlled = controlledOpen !== undefined;
+  const open = isControlled ? controlledOpen : internalOpen;
+  const setOpen = isControlled ? onOpenChange : setInternalOpen;
+
+  const utils = api.useUtils();
+
+  const integrationQuery = api.integration.listWithStats.useQuery();
+  const connection = integrationQuery.data?.active.find(
+    (int) => int.integrationId === integrationId,
+  );
+
+  const createMetric = api.metric.create.useMutation({
+    onSuccess: () => {
+      void utils.metric.getAll.invalidate();
+      setOpen?.(false);
+      setError(null);
+      onSuccess?.();
+    },
+    onError: (err) => {
+      setError(err.message);
+    },
+  });
+
+  const handleSubmit = (data: MetricCreateInput) => {
+    setError(null);
+    createMetric.mutate(data);
+  };
+
+  if (!connection) {
+    return (
+      <Card>
+        <CardHeader>
+          <CardTitle>No {integrationId} Connection</CardTitle>
+        </CardHeader>
+        <CardContent>
+          <p className="text-muted-foreground text-sm">
+            Please connect your {integrationId} account first.
+          </p>
+        </CardContent>
+      </Card>
+    );
+  }
+
+  return (
+    <Dialog open={open} onOpenChange={setOpen}>
+      {trigger && <DialogTrigger asChild>{trigger}</DialogTrigger>}
+      <DialogContent className="max-h-[80vh] overflow-y-auto sm:max-w-[600px]">
+        <DialogHeader>
+          <DialogTitle>{title}</DialogTitle>
+          {description && <DialogDescription>{description}</DialogDescription>}
+        </DialogHeader>
+
+        {children({
+          connection,
+          onSubmit: handleSubmit,
+          isCreating: createMetric.isPending,
+          error,
+        })}
+
+        {error && <p className="text-destructive text-sm">Error: {error}</p>}
+      </DialogContent>
+    </Dialog>
+  );
+}
